+++ zero-cost-templating/tests/ui/pass/working_no_loop.rs
@@ -3,8 +3,11 @@
 
 extern crate alloc;
 
+use std::pin::pin;
+
 use alloc::borrow::Cow;
 
+use futures::StreamExt;
 use zero_cost_templating_macros::template_stream;
 
 #[template_stream("test.html.hbs")]
@@ -19,4 +22,13 @@ pub async fn test() {
     template.test_copyright_year7(Cow::from("2023".to_string()));
 }
 
-pub fn main() {}
+#[tokio::main]
+pub async fn main() {
+    let mut actual = String::new();
+    let stream = test();
+    let mut stream = pin!(stream);
+    while let Some(value) = stream.next().await {
+        actual.push_str(&value);
+    }
+    assert_eq!("", actual);
+}
